+++ src/services/DemoRunner.ts
@@ -1,5 +1,5 @@
-import { PresenterView } from "./../presenterView/PresenterView";
-import { COMMAND, Config, ContextKeys, StateKeys, WebViewMessages } from "../constants";
+import { PresenterView } from './../presenterView/PresenterView';
+import { COMMAND, Config, ContextKeys, StateKeys, WebViewMessages } from '../constants';
 import {
   Action,
   Demo,
@@ -10,8 +10,8 @@ import {
   Step,
   Subscription,
   Version,
-} from "../models";
-import { Extension } from "./Extension";
+} from '../models';
+import { Extension } from './Extension';
 import {
   Position,
   Range,
@@ -26,9 +26,9 @@ import {
   commands,
   window,
   workspace,
-} from "vscode";
-import { FileProvider } from "./FileProvider";
-import { DemoPanel } from "../panels/DemoPanel";
+} from 'vscode';
+import { FileProvider } from './FileProvider';
+import { DemoPanel } from '../panels/DemoPanel';
 import {
   getVariables,
   getFileContents,
@@ -52,20 +52,21 @@ import {
   updateConfig,
   togglePresentationView,
   removeDemosForCurrentPosition,
-} from "../utils";
-import { ActionTreeItem } from "../providers/ActionTreeviewProvider";
-import { DecoratorService } from "./DecoratorService";
-import { Notifications } from "./Notifications";
-import { parse as jsonParse } from "jsonc-parser";
-import { Logger } from "./Logger";
-import { NotesService } from "./NotesService";
-import { ScriptExecutor } from "./ScriptExecutor";
-import { StateManager } from "./StateManager";
-import { Preview } from "../preview/Preview";
-import { DemoStatusBar } from "./DemoStatusBar";
+} from '../utils';
+import { ActionTreeItem } from '../providers/ActionTreeviewProvider';
+import { DecoratorService } from './DecoratorService';
+import { Notifications } from './Notifications';
+import { parse as jsonParse } from 'jsonc-parser';
+import { Logger } from './Logger';
+import { NotesService } from './NotesService';
+import { ScriptExecutor } from './ScriptExecutor';
+import { StateManager } from './StateManager';
+import { Preview } from '../preview/Preview';
+import { DemoStatusBar } from './DemoStatusBar';
+import { ExternalAppsService } from './ExternalAppsService';
 
 const DEFAULT_START_VALUE = {
-  filePath: "",
+  filePath: '',
   version: 2 as Version,
   demo: [],
 };
@@ -73,7 +74,7 @@ const DEFAULT_START_VALUE = {
 export class DemoRunner {
   private static isPresentationMode = false;
   private static terminal: { [id: string]: Terminal | null } = {};
-  private static readonly terminalName = "DemoTime";
+  private static readonly terminalName = 'DemoTime';
   private static crntFilePath: string | undefined;
   private static crntHighlightRange: Range | undefined;
   private static crntZoom: number | undefined;
@@ -88,12 +89,21 @@ export class DemoRunner {
 
     subscriptions.push(commands.registerCommand(COMMAND.start, DemoRunner.start));
     subscriptions.push(commands.registerCommand(COMMAND.previous, DemoRunner.previous));
-    subscriptions.push(commands.registerCommand(COMMAND.togglePresentationMode, DemoRunner.togglePresentationMode));
+    subscriptions.push(
+      commands.registerCommand(COMMAND.togglePresentationMode, DemoRunner.togglePresentationMode),
+    );
     subscriptions.push(commands.registerCommand(COMMAND.runStep, DemoRunner.startDemo));
     subscriptions.push(commands.registerCommand(COMMAND.runById, DemoRunner.runById));
     subscriptions.push(commands.registerCommand(COMMAND.reset, DemoRunner.reset));
-    subscriptions.push(commands.registerCommand(COMMAND.toggleHighlight, DemoRunner.toggleHighlight));
-    subscriptions.push(commands.registerCommand(COMMAND.toggleSelectionHighlight, DemoRunner.toggleSelectionHighlight));
+    subscriptions.push(
+      commands.registerCommand(COMMAND.toggleHighlight, DemoRunner.toggleHighlight),
+    );
+    subscriptions.push(
+      commands.registerCommand(
+        COMMAND.toggleSelectionHighlight,
+        DemoRunner.toggleSelectionHighlight,
+      ),
+    );
 
     window.onDidChangeActiveTextEditor(async (editor) => {
       if (editor && editor.document.fileName === DemoRunner.crntFilePath) {
@@ -115,7 +125,12 @@ export class DemoRunner {
    * @param zoom - The zoom level for the highlighting.
    * @param highlightWholeLine - Indicates whether the highlighting should be applied to the whole line.
    */
-  public static setCrntHighlighting(filePath?: string, range?: Range, zoom?: number, highlightWholeLine?: boolean) {
+  public static setCrntHighlighting(
+    filePath?: string,
+    range?: Range,
+    zoom?: number,
+    highlightWholeLine?: boolean,
+  ) {
     DemoRunner.crntFilePath = filePath;
     DemoRunner.crntHighlightRange = range;
     DemoRunner.crntZoom = zoom;
@@ -173,7 +188,9 @@ export class DemoRunner {
    * @returns {Version} The detected version of the current demo file.
    */
   public static getCurrentVersion(): Version {
-    const executingFile = Extension.getInstance().getState<DemoFileCache>(StateKeys.executingDemoFile);
+    const executingFile = Extension.getInstance().getState<DemoFileCache>(
+      StateKeys.executingDemoFile,
+    );
     if (!executingFile) {
       return 2;
     }
@@ -184,7 +201,7 @@ export class DemoRunner {
     }
 
     // Only old demo files without a version property should be version 1
-    return typeof executingFile.version === "number" ? (executingFile.version as Version) : 1;
+    return typeof executingFile.version === 'number' ? (executingFile.version as Version) : 1;
   }
 
   /**
@@ -195,12 +212,16 @@ export class DemoRunner {
    * @returns A promise that resolves when the presentation mode is toggled.
    */
   private static async togglePresentationMode(enable?: boolean): Promise<void> {
-    DemoRunner.isPresentationMode = typeof enable !== "undefined" ? enable : !DemoRunner.isPresentationMode;
+    DemoRunner.isPresentationMode =
+      typeof enable !== 'undefined' ? enable : !DemoRunner.isPresentationMode;
     DemoStatusBar.setPresenting(DemoRunner.isPresentationMode);
     await setContext(ContextKeys.presentation, DemoRunner.isPresentationMode);
-    PresenterView.postMessage(WebViewMessages.toWebview.updatePresentationStarted, DemoRunner.isPresentationMode);
+    PresenterView.postMessage(
+      WebViewMessages.toWebview.updatePresentationStarted,
+      DemoRunner.isPresentationMode,
+    );
     if (DemoRunner.isPresentationMode) {
-      DemoPanel.updateMessage("Presentation mode enabled");
+      DemoPanel.updateMessage('Presentation mode enabled');
       await DemoRunner.getDemoFile(undefined, true);
       Preview.postMessage(WebViewMessages.toWebview.updateIsInPresentationMode, true);
     } else {
@@ -231,7 +252,9 @@ export class DemoRunner {
    *
    * @returns {Promise<void>} A promise that resolves when the demo runner has started.
    */
-  private static async start(item: ActionTreeItem | { demoFilePath: string; description: string }): Promise<void> {
+  private static async start(
+    item: ActionTreeItem | { demoFilePath: string; description: string },
+  ): Promise<void> {
     if (Preview.isListening()) {
       return;
     }
@@ -247,7 +270,7 @@ export class DemoRunner {
     let demos: Demo[] = demoFile?.demo.demos || [];
 
     if (demos.length <= 0) {
-      Notifications.error("No demo steps found");
+      Notifications.error('No demo steps found');
       return;
     }
 
@@ -263,8 +286,12 @@ export class DemoRunner {
       // Check if there is a next demo file
       const nextFile = await getNextDemoFile(demoFile);
       if (!nextFile) {
-        const yesOrNo = await Notifications.info("No next demo steps found. Do you want to reset?", "Yes", "No");
-        if (yesOrNo === "Yes") {
+        const yesOrNo = await Notifications.info(
+          'No next demo steps found. Do you want to reset?',
+          'Yes',
+          'No',
+        );
+        if (yesOrNo === 'Yes') {
           await DemoRunner.reset();
           await commands.executeCommand(COMMAND.start);
         }
@@ -278,7 +305,7 @@ export class DemoRunner {
       // Start the next demo file
       DemoRunner.start({
         demoFilePath: nextFile.filePath,
-        description: nextFile.filePath.split("/").pop(),
+        description: nextFile.filePath.split('/').pop(),
       });
       return;
     }
@@ -324,7 +351,7 @@ export class DemoRunner {
     const demos = demoFile?.demos || [];
 
     if (demos.length <= 0) {
-      Notifications.error("No demo steps found");
+      Notifications.error('No demo steps found');
       return;
     }
 
@@ -339,7 +366,7 @@ export class DemoRunner {
         filePath,
       });
       if (!previousFile) {
-        Notifications.info("No previous demo steps found");
+        Notifications.info('No previous demo steps found');
         return;
       }
 
@@ -390,7 +417,11 @@ export class DemoRunner {
    * @param demoToRun - The demo to run.
    * @returns A promise that resolves when the demo execution is complete.
    */
-  private static async startDemo(demoToRun: { filePath: string; idx: number; demo: Demo }): Promise<void> {
+  private static async startDemo(demoToRun: {
+    filePath: string;
+    idx: number;
+    demo: Demo;
+  }): Promise<void> {
     if (!demoToRun) {
       return;
     }
@@ -449,7 +480,7 @@ export class DemoRunner {
     }
 
     if (!filePath) {
-      Notifications.error("No demo found with the specified id");
+      Notifications.error('No demo found with the specified id');
       return;
     }
 
@@ -462,7 +493,7 @@ export class DemoRunner {
     // Get the demo idx
     const demoIdx = demoFiles[filePath].demos.findIndex((demo) => demo.id === id);
     if (demoIdx < 0) {
-      Notifications.error("No demo found with the specified id");
+      Notifications.error('No demo found with the specified id');
       return;
     }
     const demoToRun = demoFiles[filePath].demos[demoIdx];
@@ -548,7 +579,10 @@ export class DemoRunner {
       // Verify if the current step has a `STATE_` or `SCRIPT_` variable which needs to be updated
       // This can happen when the `setState` action is used during the current demo execution (previous step)
       let stepJson = JSON.stringify(step);
-      if ((stepJson.includes(StateKeys.prefix.state) || stepJson.includes(StateKeys.prefix.script)) && variables) {
+      if (
+        (stepJson.includes(StateKeys.prefix.state) || stepJson.includes(StateKeys.prefix.script)) &&
+        variables
+      ) {
         stepJson = await insertVariables(stepJson, variables);
         step = jsonParse(stepJson);
       }
@@ -558,9 +592,19 @@ export class DemoRunner {
         await sleep(step.timeout || 1000);
         continue;
       } else if (step.action === Action.WaitForInput) {
-        const answer = await getUserInput("Press any key to continue");
+        const answer = await getUserInput('Press any key to continue');
         if (answer === undefined) {
           return;
+        }
+        continue;
+      }
+
+      // Open external applications
+      if (step.action === Action.openPowerPoint) {
+        try {
+          await ExternalAppsService.openPowerPoint();
+        } catch (error) {
+          Notifications.error(`Failed to open PowerPoint: ${(error as Error).message}`);
         }
         continue;
       }
@@ -568,7 +612,7 @@ export class DemoRunner {
       // Update settings
       if (step.action === Action.SetSetting) {
         if (!step.setting || !step.setting.key) {
-          Notifications.error("No setting key or value specified");
+          Notifications.error('No setting key or value specified');
           continue;
         }
 
@@ -579,16 +623,16 @@ export class DemoRunner {
 
       if (step.action === Action.SetTheme) {
         if (!step.theme) {
-          Notifications.error("No theme specified");
+          Notifications.error('No theme specified');
           continue;
         }
 
-        await updateConfig("workbench.colorTheme", step.theme);
+        await updateConfig('workbench.colorTheme', step.theme);
         continue;
       }
 
       if (step.action === Action.UnsetTheme) {
-        await updateConfig("workbench.colorTheme", null);
+        await updateConfig('workbench.colorTheme', null);
         continue;
       }
 
@@ -605,7 +649,7 @@ export class DemoRunner {
       // Set state
       if (step.action === Action.SetState) {
         if (!step.state || !step.state.key || !step.state.value) {
-          Notifications.error("No state key or value specified");
+          Notifications.error('No state key or value specified');
           return;
         }
 
@@ -619,7 +663,7 @@ export class DemoRunner {
       // Execute the specified VSCode command
       if (step.action === Action.ExecuteVSCodeCommand) {
         if (!step.command) {
-          Notifications.error("No command specified");
+          Notifications.error('No command specified');
           continue;
         }
 
@@ -634,7 +678,7 @@ export class DemoRunner {
 
       if (step.action === Action.ShowInfoMessage) {
         if (!step.message) {
-          Notifications.error("No message specified");
+          Notifications.error('No message specified');
           continue;
         }
 
@@ -644,14 +688,14 @@ export class DemoRunner {
 
       if (step.action === Action.OpenWebsite) {
         if (!step.url) {
-          Notifications.error("No URL specified");
+          Notifications.error('No URL specified');
           continue;
         }
 
         if (step.openInVSCode) {
-          await commands.executeCommand("simpleBrowser.show", Uri.parse(step.url));
+          await commands.executeCommand('simpleBrowser.show', Uri.parse(step.url));
         } else {
-          await commands.executeCommand("vscode.open", Uri.parse(step.url));
+          await commands.executeCommand('vscode.open', Uri.parse(step.url));
           continue;
         }
       }
@@ -678,15 +722,15 @@ export class DemoRunner {
         // Write the content at the current position
         const editor = window.activeTextEditor;
         if (!editor) {
-          Notifications.error("No active text editor found");
+          Notifications.error('No active text editor found');
           return;
         }
 
         const position = editor.selection.active;
-        const content = step.content || "";
+        const content = step.content || '';
 
         if (!content) {
-          Notifications.error("No content to write");
+          Notifications.error('No content to write');
           return;
         }
 
@@ -695,7 +739,7 @@ export class DemoRunner {
       }
 
       if (step.action === Action.Format) {
-        await commands.executeCommand("editor.action.formatDocument");
+        await commands.executeCommand('editor.action.formatDocument');
         continue;
       }
 
@@ -704,12 +748,12 @@ export class DemoRunner {
       }
 
       if (step.action === Action.Close) {
-        await commands.executeCommand("workbench.action.closeActiveEditor");
+        await commands.executeCommand('workbench.action.closeActiveEditor');
         continue;
       }
 
       if (step.action === Action.CloseAll) {
-        await commands.executeCommand("workbench.action.closeAllEditors");
+        await commands.executeCommand('workbench.action.closeAllEditors');
         continue;
       }
 
@@ -733,12 +777,12 @@ export class DemoRunner {
       }
 
       if (step.action === Action.Open) {
-        await commands.executeCommand("vscode.open", fileUri);
+        await commands.executeCommand('vscode.open', fileUri);
         continue;
       }
 
       if (step.action === Action.MarkdownPreview) {
-        await commands.executeCommand("markdown.showPreview", fileUri);
+        await commands.executeCommand('markdown.showPreview', fileUri);
         continue;
       }
 
@@ -762,7 +806,7 @@ export class DemoRunner {
         continue;
       }
 
-      let content = step.content || "";
+      let content = step.content || '';
       if (step.contentPath) {
         const fileContent = await getFileContents(workspaceFolder, step.contentPath);
         if (!fileContent) {
@@ -794,27 +838,41 @@ export class DemoRunner {
       if (step.action === Action.Highlight && (crntRange || crntPosition)) {
         let highlightWholeLine = step.highlightWholeLine;
         if (usesPlaceholders) {
-          highlightWholeLine = typeof step.highlightWholeLine === "undefined" ? true : step.highlightWholeLine;
+          highlightWholeLine =
+            typeof step.highlightWholeLine === 'undefined' ? true : step.highlightWholeLine;
         }
 
-        await DemoRunner.highlight(textEditor, crntRange, crntPosition, step.zoom, highlightWholeLine);
+        await DemoRunner.highlight(
+          textEditor,
+          crntRange,
+          crntPosition,
+          step.zoom,
+          highlightWholeLine,
+        );
         continue;
       }
 
       // Code actions
       if (step.action === Action.Insert) {
-        await DemoRunner.insert(textEditor, editor, fileUri, content, crntPosition, step.lineInsertionDelay);
+        await DemoRunner.insert(
+          textEditor,
+          editor,
+          fileUri,
+          content,
+          crntPosition,
+          step.lineInsertionDelay,
+        );
         continue;
       }
 
       if (step.action === Action.Write) {
         if (!content) {
-          Notifications.error("No content to write");
+          Notifications.error('No content to write');
           return;
         }
 
         if (!crntPosition) {
-          Notifications.error("No position specified where to write the content");
+          Notifications.error('No position specified where to write the content');
           return;
         }
 
@@ -830,14 +888,17 @@ export class DemoRunner {
           content,
           crntRange,
           crntPosition,
-          step.lineInsertionDelay
+          step.lineInsertionDelay,
         );
         continue;
       }
 
       if (step.action === Action.PositionCursor) {
         if (crntPosition) {
-          textEditor.revealRange(new Range(crntPosition, crntPosition), TextEditorRevealType.InCenter);
+          textEditor.revealRange(
+            new Range(crntPosition, crntPosition),
+            TextEditorRevealType.InCenter,
+          );
           textEditor.selection = new Selection(crntPosition, crntPosition);
         }
         continue;
@@ -868,7 +929,7 @@ export class DemoRunner {
     fileUri: Uri,
     content: string,
     position: Position | undefined,
-    lineInsertionDelay?: number
+    lineInsertionDelay?: number,
   ): Promise<void> {
     if (!position) {
       return;
@@ -893,7 +954,7 @@ export class DemoRunner {
       } else {
         const lineRange = getLineRange(editor, position);
         if (!lineRange) {
-          Logger.error("Line range not found");
+          Logger.error('Line range not found');
           return;
         }
         textEditor.revealRange(lineRange, TextEditorRevealType.InCenter);
@@ -907,10 +968,10 @@ export class DemoRunner {
       } else {
         const range = getLineRange(editor, position);
         if (!range) {
-          Logger.error("Line range not found");
+          Logger.error('Line range not found');
           return;
         }
-        await replaceContent(fileUri, range, "");
+        await replaceContent(fileUri, range, '');
         textEditor.revealRange(range, TextEditorRevealType.InCenter);
         await insertLineByLine(fileUri, range.start.line, content, lineSpeed);
       }
@@ -942,7 +1003,7 @@ export class DemoRunner {
     content: string,
     range: Range | undefined,
     position: Position | undefined,
-    lineInsertionDelay?: number
+    lineInsertionDelay?: number,
   ): Promise<void> {
     if (!range && !position) {
       return;
@@ -959,7 +1020,7 @@ export class DemoRunner {
         const start = new Position(startLine.lineNumber, 0);
         const end = new Position(endLine.lineNumber, endLine.text.length);
 
-        await replaceContent(fileUri, new Range(start, end), "");
+        await replaceContent(fileUri, new Range(start, end), '');
 
         textEditor.revealRange(new Range(start, end), TextEditorRevealType.InCenter);
         await insertLineByLine(fileUri, startLine.lineNumber, content, lineSpeed);
@@ -974,11 +1035,11 @@ export class DemoRunner {
         range = getLineRange(editor, position);
 
         if (!range) {
-          Logger.error("Line range not found");
+          Logger.error('Line range not found');
           return;
         }
 
-        await replaceContent(fileUri, range, "");
+        await replaceContent(fileUri, range, '');
 
         textEditor.revealRange(range, TextEditorRevealType.InCenter);
         await insertLineByLine(fileUri, range.start.line, content, lineSpeed);
@@ -1006,7 +1067,7 @@ export class DemoRunner {
     editor: TextDocument,
     fileUri: Uri,
     range: Range | undefined,
-    position: Position | undefined
+    position: Position | undefined,
   ): Promise<void> {
     if (!range && !position) {
       return;
@@ -1077,7 +1138,8 @@ export class DemoRunner {
     const selection = activeEditor.selection;
     const range = selection.isSingleLine ? selection : new Range(selection.start, selection.end);
     const endLineText = activeEditor.document.lineAt(selection.end).text;
-    let highlightWholeLine = range.start.character === 0 && range.end.character === endLineText.length;
+    let highlightWholeLine =
+      range.start.character === 0 && range.end.character === endLineText.length;
     if (range.start.line === range.end.line && range.start.character === range.end.character) {
       highlightWholeLine = true;
     }
@@ -1104,7 +1166,7 @@ export class DemoRunner {
     position: Position | undefined,
     zoomLevel?: number,
     highlightWholeLine?: boolean,
-    keepInMemory = true
+    keepInMemory = true,
   ): Promise<void> {
     if (!range && !position) {
       return;
@@ -1120,7 +1182,12 @@ export class DemoRunner {
 
     if (range) {
       if (keepInMemory) {
-        DemoRunner.setCrntHighlighting(textEditor.document.fileName, range, zoomLevel, highlightWholeLine);
+        DemoRunner.setCrntHighlighting(
+          textEditor.document.fileName,
+          range,
+          zoomLevel,
+          highlightWholeLine,
+        );
         await setContext(ContextKeys.hasCodeHighlighting, true);
       }
 
@@ -1142,9 +1209,12 @@ export class DemoRunner {
    * @param command - The command to be executed.
    * @returns A promise that resolves when the command execution is complete.
    */
-  private static async executeTerminalCommand(command?: string, terminalId?: string): Promise<void> {
+  private static async executeTerminalCommand(
+    command?: string,
+    terminalId?: string,
+  ): Promise<void> {
     if (!command) {
-      Notifications.error("No command specified");
+      Notifications.error('No command specified');
       return;
     }
 
@@ -1179,7 +1249,7 @@ export class DemoRunner {
       return;
     }
 
-    terminal.sendText("exit", true);
+    terminal.sendText('exit', true);
     terminal.dispose();
     delete DemoRunner.terminal[terminalId];
   }
@@ -1194,9 +1264,13 @@ export class DemoRunner {
    *
    * @throws Will throw an error if the destination is not specified or if the rename operation fails.
    */
-  private static async rename(workspaceFolder: WorkspaceFolder, fileUri: Uri, step: Step): Promise<void> {
+  private static async rename(
+    workspaceFolder: WorkspaceFolder,
+    fileUri: Uri,
+    step: Step,
+  ): Promise<void> {
     if (!step.dest) {
-      Notifications.error("No destination specified");
+      Notifications.error('No destination specified');
       return;
     }
 
@@ -1217,9 +1291,13 @@ export class DemoRunner {
    * @returns A promise that resolves when the copy operation is complete.
    * @throws Will throw an error if the destination is not specified or if the copy operation fails.
    */
-  private static async copy(workspaceFolder: WorkspaceFolder, fileUri: Uri, step: Step): Promise<void> {
+  private static async copy(
+    workspaceFolder: WorkspaceFolder,
+    fileUri: Uri,
+    step: Step,
+  ): Promise<void> {
     if (!step.dest) {
-      Notifications.error("No destination specified");
+      Notifications.error('No destination specified');
       return;
     }
 
@@ -1252,7 +1330,7 @@ export class DemoRunner {
    * @returns A promise that resolves when the file is saved.
    */
   private static async saveFile(): Promise<void> {
-    await commands.executeCommand("workbench.action.files.save");
+    await commands.executeCommand('workbench.action.files.save');
   }
 
   /**
@@ -1263,7 +1341,7 @@ export class DemoRunner {
    */
   private static async getDemoFile(
     item?: ActionTreeItem | Uri,
-    triggerFirstDemo: boolean = false
+    triggerFirstDemo: boolean = false,
   ): Promise<
     | {
         filePath: string;
@@ -1278,13 +1356,13 @@ export class DemoRunner {
 
     if (item && itemPath) {
       if (!demoFiles) {
-        Notifications.warning("No demo files found");
+        Notifications.warning('No demo files found');
         return;
       }
 
       const demoFile = await FileProvider.getFile(Uri.file(itemPath));
       if (!demoFile) {
-        const demoFileName = itemPath.split("/").pop();
+        const demoFileName = itemPath.split('/').pop();
         Notifications.warning(`No demo file found with the name ${demoFileName}`);
         return;
       }
@@ -1326,7 +1404,7 @@ export class DemoRunner {
     } else if (executingFile.filePath && !item && demoFiles) {
       const demoFile = demoFiles[executingFile.filePath];
       if (!demoFile) {
-        Notifications.warning("No demo file found");
+        Notifications.warning('No demo file found');
         return;
       }